rockchip: firefly: Save the environment on SD card
Save the environment on the SD card for Firefly in the empty space
between the SPL and the u-boot image.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>

Cover-letter:
Small Configuration tweaks for rockchip/firefly

Further testing with u-boot on Rockchip revealed that loading the initrd
above 512M seems to be problematic (just like with the dtb) and ofcourse
that not being able to save the environment is annoying :)

Resolve these things.
END
Series-to: Simon Glass <sjg@chromium.org>