rockchip: Drop Ethernet from the TODO
Now that ethernet support works, it can be dropped from the rockchip
TODO

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>

Series-to: Simon Glass <sjg@chromium.org>
Series-cc: Heiko Stuebner <heiko@sntech.de>
Series-cc: Hans de Goede <hdegoede@redhat.com>

Cover-letter:
Add support for Rockchip RK3288 Ethernet

This series adds support for the GMAC Ethernet interface on RK3288 SoCs.

To add support I've taken a slightly different approach then some of the
other boards with a designware IP block, by creating a new driver to
take care of the platfrom glue which subclasses the main designware driver
instead of adding the compatibility string the designware driver
directly and doing the SoC specific setup in the board files. This seems
quite a bit more elegant in a device model based world.

As the pinctrl and clock drivers are quite simple at the moment
(hardcoded settings rather then retrieved from device-tree) the pinctrl
and clock settings added in this series assume the setup for the
ethernet interface used is for Firefly and Radxa Rock 2. Specifically It
assumes the gmac is driven by an external clock, GPIO4B0 is the phy
reset and the phy interface mode is RGMII.

I've only tested this series on a Radxa Rock 2 board, it would be great
if someone could test this on other boards with the designware IP
especially for those with the reset GPIO in devicetree (e.g. some of the
Allwinner boards).

END